BUGFIX Sometimes the folders get when deleting draft (never published) pages

+++ code/ImageGalleryAlbum.php
@@ -75,7 +75,7 @@ class ImageGalleryAlbum extends DataObject
 	{
 		parent::onBeforeDelete();
 		$this->GalleryItems()->removeAll();
-		$this->Folder()->delete();
+		if($this->Folder()->ID) $this->Folder()->delete();
 	}
